requirements.txt
alembic==0.9.5
aniso8601==1.3.0
appnope==0.1.0
bcrypt==3.1.4
cffi==1.11.2
click==6.7
decorator==4.1.2
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
ipdb==0.10.3
ipython==6.2.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.6
marshmallow-sqlalchemy==0.13.1
nose==1.3.7
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.14
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
WTForms==2.1
